+++ docs/history.md
@@ -11,6 +11,7 @@
 - The second phase of Mega Satan no longer has armor.
 - Marrow, Tropicamide, and Dogma can no longer be given as a random item during diversity races.
 - Fixed the bug where the diversity documentation was not accurate with respect to the excluded starting items.
+- Fixed the bug where it was possible to reroll into placeholder items.
 
 ### _v0.82.13_ - August 14th, 2022